refactor: apply new folder structure
