+++ src/app/constructor-standings/constructor-standings.component.scss
