+++ src/app/driver-standings/utils/driver-standing.interface.ts
@@ -1,0 +1,10 @@
+export interface DriverStanding {
+  firstName: string;
+  lastName: string;
+  driverId: string;
+  constructorName: string;
+  constructorId: string;
+  points: string;
+  position: number;
+  driverNumber: string;
+}
