+++ src/app/shared/interfaces/standings.interface.ts
